pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-bio-experiment"
version = "0.1.0"
description = "RL environment for biological experiment pipeline planning"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
train = [
    "gymnasium>=0.29.0",
]
bio = [
    "biopython>=1.84",
    "gseapy>=1.1.3",
    "scanpy>=1.10.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "gymnasium>=0.29.0",
]

[project.scripts]
server = "hackathon.server.app:main"

[tool.uv]
package = false

[tool.setuptools]
include-package-data = true
packages = [
    "hackathon",
    "hackathon.server",
    "hackathon.server.simulator",
    "hackathon.server.rules",
    "hackathon.server.rewards",
    "hackathon.server.tasks",
    "hackathon.server.subagents",
    "hackathon.training",
    "hackathon.tests",
]
[tool.setuptools.package-dir]
hackathon = "."
"hackathon.server" = "server"
"hackathon.server.simulator" = "server/simulator"
"hackathon.server.rules" = "server/rules"
"hackathon.server.rewards" = "server/rewards"
"hackathon.server.tasks" = "server/tasks"
"hackathon.server.subagents" = "server/subagents"
"hackathon.training" = "training"
"hackathon.tests" = "tests"
